src/requirements.txt
grpcio
protobuf
pydantic
sqlalchemy
psycopg2-binary
alembic